+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
+- support rendering in a repeated MapView
 - animate internally so that deck.gl _animate is not necessary anymore
+- add imageUnscale param to allow arbitrary scaled data
 
 ## 1.0.9